requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
black==20.8b1
cffi==1.14.5
click==7.1.2
dataclasses==0.6
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonify==0.5
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
pycparser==2.20
regex==2021.3.17
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
